Exclude js files from type tree
In order to allow editors to resolve modules it is necessary to set
additional paths in `tsconfig.json`. These files leak past the
TypeScript source tree via the type tree and prevent the TypeScript
Compiler from writing the compiled files as a file already exists
with the same name. This change adds a filter on the type tree to
stop this leak