Use ember-try to hack around qunit nodetests issue.
Add an ember:try configuration for running nodetests. In it,
remove the ember-cli-qunit dependency from the project, so that
the generator test can successfully run with its faked-out
dependency instead of finding the *actual* project dependency.

Update Travis and AppVeyor configs to use the ember-try scenario
instead of running `nodetest` directly.