Add resolution for @types/ember.
We need to make sure that the "latest" dependency added
implicitly by `@types/ember-qunit` doesn't cause duplicate
type installations when yarn tries (sensibly in the normal
scenario) to keep existing versions in place. That's reasonable
as a solution to transitive dependency hell; it's just a poor fit
for TS, which cannot resolve multiple copies of the same type
declarations if installed.