Update tsconfig and corresponding docs
Our generated blueprint was *years* out of date, and between the
changes we made in the types over the last year or so and the guidance
we have given in the TS RFCs we are now both *able* and *required* to
set `strict: true` as well as to opt into `noUncheckedIndexedAccess`
and to disable `allowSyntheticDefaultImports` and `esModuleInterop`.

Additionally, this provided a nice opportunity to document the choices
we have made so folks have something to latch onto when looking at the
config we generate for them, and to update the config to target the
current latest versions of the `target` and `module` fields.