docs: fix typos and replace use of attr()
Currently in the docs, `@attr() declare name?` is typed as `string`, and no example of
`@attr('string')` is given. I'm not sure if showing in the documentation
that not passing a transform to `@attr()` should explicitly default to
string is the correct thing to do, but I definitely think we should
model the use of `@attr('string')` in the code examples.