docs: Overhaul component documentation to use Signatures
The current documentation on components only discusses the arguments
passed to a Glimmer component. Given RFC 0748's implementation of a
`Signature` type with `Element`, `Args`, and `Blocks` fields, I think
it's appropriate to update the component docs accordingly.

The earlier version of the page spent a lot of time discussing the
minutiae of how `this.args` comes to be typed, etc., and focused on the
component's `constructor`. Both seem to me a little off-track for someone
eager to write an Octane app in TypeScript, especially given how native
class syntax intercepts the need for a constructor.

I've instead broken up the document into a section with no signature,
the signature itself, args, blocks, element, and then, finally, how to
subclass one's own component.

A lot of the text I've adapted/copied from either RFC 0748 or written
based on advice (esp wrt subclassing) I've gotten from Chris on Discord.
I think it's a good refresh of documentation that is a few years old.
It's also looking forward, I hope, to landing some of this documentation
in Ember proper.

If the discussion of constructors and `owner` is still important (I'm
basing this on my personal experience), then we can add it back in, but
I'd need help updating that section, I think.