Use `paths` instead of `typeRoots` for `ts` project
Technically `typeRoots` *works*, but we don't generally want people
doing that, and our tsconfig files get used as references. `paths` does
the same thing in a way that *does* match how we want people to work,
at least until such a time as we're in a more normal module resolution.